+++ src/Opravilo.API/Models/Responses/Account/AuthenticationResponse.cs
@@ -1,11 +1,16 @@
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 
-namespace Opravilo.API.Models.Responses
+namespace Opravilo.API.Models.Responses.Account
 {
     public class AuthenticationResponse
     {
+        [Required]
         public string Token { get; set; }
+        
+        [Required]
         public bool IsSuccess { get; set; }
+        
         public List<string> Errors { get; set; }
     }
 }
